modifying hacc analysis to latest standard
